+++ source/renderer/app/stores/NetworkStatusStore.js
@@ -353,12 +353,11 @@ export default class NetworkStatusStore extends Store {
     const wasConnected = this.isConnected;
 
     try {
-      let networkStatus: GetNetworkStatusResponse;
-      if (isForcedTimeDifferenceCheck) {
-        networkStatus = await this.forceCheckTimeDifferenceRequest.execute(queryParams).promise;
-      } else {
-        networkStatus = await this.getNetworkStatusRequest.execute().promise;
-      }
+      const networkStatus: GetNetworkStatusResponse = isForcedTimeDifferenceCheck ? (
+        await this.forceCheckTimeDifferenceRequest.execute(queryParams).promise
+      ) : (
+        await this.getNetworkStatusRequest.execute().promise
+      );
 
       // In case we no longer have TLS config we ignore all API call responses
       // as this means we are in the Cardano shutdown (stopping|exiting|updating) sequence
@@ -372,7 +371,7 @@ export default class NetworkStatusStore extends Store {
         syncProgress,
         blockchainHeight,
         localBlockchainHeight,
-        localTimeDifference,
+        localTimeInformation,
       } = networkStatus;
 
       // We got response which means node is responding
@@ -388,11 +387,13 @@ export default class NetworkStatusStore extends Store {
 
       // System time is correct if local time difference is below allowed threshold
       runInAction('update localTimeDifference and isNodeTimeCorrect', () => {
-        this.localTimeDifference = localTimeDifference;
-        this.isNodeTimeCorrect = (
-          this.localTimeDifference !== null && // If we receive 'null' it means NTP check failed
-          this.localTimeDifference <= ALLOWED_TIME_DIFFERENCE
-        );
+        if (localTimeInformation.status !== 'pending') {
+          this.localTimeDifference = localTimeInformation.difference;
+          this.isNodeTimeCorrect = (
+            this.localTimeDifference != null && // If we receive 'null' it means NTP check failed
+            this.localTimeDifference <= ALLOWED_TIME_DIFFERENCE
+          );
+        }
       });
 
       if (this._networkStatus === NETWORK_STATUS.CONNECTING && this.isNodeSubscribed) {